Move the FDB client dependencies into their own sub-package.
