+++ Dockerfile
@@ -48,6 +48,7 @@ COPY main.go main.go
 COPY api/ api/
 COPY controllers/ controllers/
 COPY setup/ setup/
+COPY fdbclient/ fdbclient/
 
 # Build
 RUN CGO_ENABLED=1 GOOS=linux GOARCH=amd64 GO111MODULE=on go build -a -o manager main.go
