+++ controllers/bounce_processes.go
@@ -37,7 +37,7 @@ type BounceProcesses struct{}
 
 // Reconcile runs the reconciler's work.
 func (b BounceProcesses) Reconcile(r *FoundationDBClusterReconciler, context ctx.Context, cluster *fdbtypes.FoundationDBCluster) (bool, error) {
-	adminClient, err := r.AdminClientProvider(cluster, r)
+	adminClient, err := r.getDatabaseClientProvider().GetAdminClient(cluster, r)
 	if err != nil {
 		return false, err
 	}
