+++ controllers/controllers.go
@@ -29,12 +29,8 @@ import (
 
 var log = logf.Log.WithName("controller")
 
-// DefaultCLITimeout is the default timeout for CLI commands.
-var DefaultCLITimeout = 10
-
 const (
 	clusterFileKey = "cluster-file"
-)
 
 // metadataMatches determines if the current metadata on an object matches the
 // metadata specified by the cluster spec.
