+++ controllers/update_database_configuration.go
@@ -37,7 +37,7 @@ type UpdateDatabaseConfiguration struct{}
 
 // Reconcile runs the reconciler's work.
 func (u UpdateDatabaseConfiguration) Reconcile(r *FoundationDBClusterReconciler, context ctx.Context, cluster *fdbtypes.FoundationDBCluster) (bool, error) {
-	adminClient, err := r.AdminClientProvider(cluster, r)
+	adminClient, err := r.getDatabaseClientProvider().GetAdminClient(cluster, r)
 
 	if err != nil {
 		return false, err
