+++ fdbclient/doc.go
@@ -1,0 +1,3 @@
+// Package fdbclient provides code for interacting with a real FoundationDB
+// cluster.
+package fdbclient
